# GS: Get rid of extra binding for channel shuffle