--- a/pcsx2/GS/Renderers/HW/GSRendererNew.h
+++ b/pcsx2/GS/Renderers/HW/GSRendererNew.h
@@ -27,7 +27,7 @@ private:
 	inline void ResetStates();
 	inline void SetupIA(const float& sx, const float& sy);
 	inline void EmulateTextureShuffleAndFbmask();
-	inline void EmulateChannelShuffle(GSTexture** rt, const GSTextureCache::Source* tex);
+	inline void EmulateChannelShuffle(const GSTextureCache::Source* tex);
 	inline void EmulateBlending(bool& DATE_PRIMID, bool& DATE_BARRIER);
 	inline void EmulateTextureSampler(const GSTextureCache::Source* tex);
 	inline void EmulateZbuffer();