### Description of Changes

Having this binding was redundant, as there's no "normal" texture sampled when we're doing a channel shuffle, and it caused issues in Vulkan when the render target or depth buffer the source.

Also fixes the Urban Chaos HLE shader.

Fixes validation errors in GT4, NFS: Carbon, Urban Chaos, probably others too.

For DX11, does a smaller copy for the RT when it's copied for sampling, which is a significant improvement (9fps->25fps with ultra blending in xenosaga, 3fps->25fps at 4x).

### Suggested Testing Steps

Test blending and channel shuffle games.