@refractionpcsx2 2022-01-31
@Dreadmoth is that all renderers? or just DX?