@Dreadmoth 2022-01-31
@refractionpcsx2  All renderers.