@stenzek 2022-02-01
Should be fixed with https://github.com/PCSX2/pcsx2/pull/5447