@web1018 2022-02-02
This build has apperantly broken the menu of Call of duty World at War Final Fronts.
 Testing on both Opengl, or Vulkan with default settings it shows black screen. On the previous build pcsx2-v1.7.2283, it is working perfectly with the same default options on Opengl and Vulkan.

You can find the image comparison below

pcsx2-v1.7.2283 and earlier
![test-working](https://user-images.githubusercontent.com/51851411/152142528-3cd0c24e-9a0e-477d-b3f7-39051760f27f.jpg)


pcsx2- 1.7.2284 (#5422) and onwards
![test-broken](https://user-images.githubusercontent.com/51851411/152142542-d9cb62b1-e211-452c-a216-db51cfee3bb5.jpg)


If you need me to open a new ticket for this to be reported, please let me know.