@lightningterror 2022-02-02
@web1018 Can you post a gsdump?