@web1018 2022-02-02
> @web1018 Can you post a gsdump?

I have never done gsdump before. I followed the instructions I found on pcsx2. If I did it incorrectly please let me know.

[COD-WAW-Final Fronts-Gsdump.zip](https://github.com/PCSX2/pcsx2/files/7986008/COD-WAW-Final.Fronts-Gsdump.zip)