@refractionpcsx2 2022-02-02
Did you extract the resources folder when you updated? I'm running your dump on the latest master and vulkan is working fine.