@web1018 2022-02-02
Oh no I havent. I didnt realize that also needed to be changed when updating. I thought its just pcsx2x64-avx2 file itself. 

Which files do I need to change when updating to new version of the build?