@refractionpcsx2 2022-02-02
It needs to be updated with *every* build, always extract everything.