@web1018 2022-02-02
Alright. Thanks for letting me know. Ignore my post above then about the issue. Im still quite new to pcsx2 emulation.