@lightningterror 2022-04-24
Adding some perf examples now, will be used for the charts for q1.
D3D11 perf charts/comparison:
Ultra/maximum blending is used for consistency but gains are also present on basic blending as well.

Resident Evil Outbreak:
Master: 120 fps
PR: 555 fps

Monster Hunter:
Master: 260 fps
PR: 620 fps

Super Dragon Ball Z
Master: 7 fps
PR: 92 fps

Fifa Street:
Master: 104 fps
PR: 133 fps

Sly 2:
Master: 17 fps
PR: 83 fps

Echo Night 
Master: 23 fps
PR: 27 fps

Some examples on basic blending:
Echo Night 
Master: 26 fps
PR: 27 fps

Fifa Street:
Master: 120 fps
PR: 134 fps